VRChat Map Maker : Unity Importer
So this is the beginning of the Unity Importer.
I'd like to also create a Blender, Godot Engine and
Unreal Engine importer, but let's focus on what people
use for VRChat first !

The project is "working", but it's still very clunky,
since you have to remap everything as its setup on
the Map Maker, so I'll see if I can automate the setup
process.

Signed-off-by: Voyage <voyage@miouyouyou.fr>